% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{apply_volume_scenario}
\alias{apply_volume_scenario}
\title{Rescale a forecast by an assumed volume change}
\usage{
apply_volume_scenario(predicted, volume_change = 0)
}
\arguments{
\item{predicted}{Numeric predictions.}

\item{volume_change}{Fractional change. A value of 0.05 means 5 percent more volume.}
}
\value{
A \code{golfops_scenario} list whose totals are the assumed counts.
}
\description{
Rescale a forecast by an assumed volume change
}
