% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golfops-package.R
\name{dobson_column_names}
\alias{dobson_column_names}
\title{Column names in the City of Mesa Dobson Ranch export}
\usage{
dobson_column_names()
}
\value{
A character vector of the 23 catalog field names, in catalog order.
}
\description{
Column names in the City of Mesa Dobson Ranch export
}
\examples{
dobson_column_names()
}
