% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golfops-package.R
\name{dobson_rounds_example}
\alias{dobson_rounds_example}
\title{Path to the bundled Dobson Ranch example extract}
\usage{
dobson_rounds_example()
}
\value{
Absolute path to the CSV shipped in \code{inst/extdata}.
}
\description{
Path to the bundled Dobson Ranch example extract
}
\examples{
dobson_rounds_example()
}
