% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golfops-package.R
\docType{package}
\name{golfops-package}
\alias{golfops}
\alias{golfops-package}
\title{golfops: Golf Participation and Revenue Summaries}
\description{
Summarizes municipal golf participation, daylight-estimated utilization, and revenue for City of Mesa Dobson Ranch. Operating charts use one documented daily sample. Revenue ratios are historical relationships, and tournament results are break-even baselines. Forecasts of observed rounds are not latent demand and not price elasticities. A labeled synthetic ticket series remains a secondary example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://data.mesaaz.gov/Parks-Recreation-and-Community-Facilities/Dobson-Ranch-Golf-Rounds/24g5-hb8k}
}

}
\author{
\strong{Maintainer}: Justin Locke \email{justinrlocke@gmail.com}

Authors:
\itemize{
  \item Justin Locke \email{justinrlocke@gmail.com}
}

}
\keyword{internal}
